Reorganize file structure and paths
