+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/bailey/turing/3mod/projects/rancid-tomatillos/src/index.js":"1","/Users/bailey/turing/3mod/projects/rancid-tomatillos/src/reportWebVitals.js":"2","/Users/bailey/turing/3mod/projects/rancid-tomatillos/src/App.js":"3","/Users/bailey/turing/3mod/projects/rancid-tomatillos/src/movieData.js":"4","/Users/bailey/turing/3mod/projects/rancid-tomatillos/src/components/ThumbnailContainer.js":"5","/Users/bailey/turing/3mod/projects/rancid-tomatillos/src/components/Header.js":"6","/Users/bailey/turing/3mod/projects/rancid-tomatillos/src/components/Thumbnail.js":"7","/Users/bailey/turing/3mod/projects/rancid-tomatillos/src/components/Movie.js":"8"},{"size":500,"mtime":1606867415473,"results":"9","hashOfConfig":"10"},{"size":362,"mtime":1606867415473,"results":"11","hashOfConfig":"10"},{"size":1105,"mtime":1606954163210,"results":"12","hashOfConfig":"10"},{"size":12953,"mtime":1606947512322,"results":"13","hashOfConfig":"10"},{"size":686,"mtime":1606950096345,"results":"14","hashOfConfig":"10"},{"size":261,"mtime":1606954312749,"results":"15","hashOfConfig":"10"},{"size":420,"mtime":1606950136791,"results":"16","hashOfConfig":"10"},{"size":622,"mtime":1606971096321,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"1tey2fc",{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/bailey/turing/3mod/projects/rancid-tomatillos/src/index.js",[],["35","36"],"/Users/bailey/turing/3mod/projects/rancid-tomatillos/src/reportWebVitals.js",[],"/Users/bailey/turing/3mod/projects/rancid-tomatillos/src/App.js",[],"/Users/bailey/turing/3mod/projects/rancid-tomatillos/src/movieData.js",[],"/Users/bailey/turing/3mod/projects/rancid-tomatillos/src/components/ThumbnailContainer.js",[],"/Users/bailey/turing/3mod/projects/rancid-tomatillos/src/components/Header.js",[],"/Users/bailey/turing/3mod/projects/rancid-tomatillos/src/components/Thumbnail.js",[],"/Users/bailey/turing/3mod/projects/rancid-tomatillos/src/components/Movie.js",[],{"ruleId":"37","replacedBy":"38"},{"ruleId":"39","replacedBy":"40"},"no-native-reassign",["41"],"no-negated-in-lhs",["42"],"no-global-assign","no-unsafe-negation"]
+[{"/Users/wolf/turing/projects/3_mod/rancid-tomatillos/src/index.js":"1","/Users/wolf/turing/projects/3_mod/rancid-tomatillos/src/reportWebVitals.js":"2","/Users/wolf/turing/projects/3_mod/rancid-tomatillos/src/App/App.js":"3","/Users/wolf/turing/projects/3_mod/rancid-tomatillos/src/movieData.js":"4","/Users/wolf/turing/projects/3_mod/rancid-tomatillos/src/Movie/Movie.js":"5","/Users/wolf/turing/projects/3_mod/rancid-tomatillos/src/Header/Header.js":"6","/Users/wolf/turing/projects/3_mod/rancid-tomatillos/src/Thumbnail-Container/ThumbnailContainer.js":"7","/Users/wolf/turing/projects/3_mod/rancid-tomatillos/src/Thumbnail/Thumbnail.js":"8"},{"size":507,"mtime":1607036394935,"results":"9","hashOfConfig":"10"},{"size":362,"mtime":1607035861984,"results":"11","hashOfConfig":"10"},{"size":1108,"mtime":1607036216354,"results":"12","hashOfConfig":"10"},{"size":12953,"mtime":1607035861983,"results":"13","hashOfConfig":"10"},{"size":618,"mtime":1607036140393,"results":"14","hashOfConfig":"10"},{"size":256,"mtime":1607036131647,"results":"15","hashOfConfig":"10"},{"size":692,"mtime":1607036549225,"results":"16","hashOfConfig":"10"},{"size":414,"mtime":1607036149057,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1wossnz",{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/wolf/turing/projects/3_mod/rancid-tomatillos/src/index.js",[],"/Users/wolf/turing/projects/3_mod/rancid-tomatillos/src/reportWebVitals.js",[],"/Users/wolf/turing/projects/3_mod/rancid-tomatillos/src/App/App.js",[],"/Users/wolf/turing/projects/3_mod/rancid-tomatillos/src/movieData.js",[],"/Users/wolf/turing/projects/3_mod/rancid-tomatillos/src/Movie/Movie.js",[],"/Users/wolf/turing/projects/3_mod/rancid-tomatillos/src/Header/Header.js",[],"/Users/wolf/turing/projects/3_mod/rancid-tomatillos/src/Thumbnail-Container/ThumbnailContainer.js",[],"/Users/wolf/turing/projects/3_mod/rancid-tomatillos/src/Thumbnail/Thumbnail.js",[]]
