+++ src/App/App.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
-import ThumbnailContainer from './components/ThumbnailContainer.js'
-import Header from './components/Header.js';
-import Movie from './components/Movie.js';
-import movieData from './movieData.js';
-import './css/App.scss';
+import ThumbnailContainer from '../Thumbnail-Container/ThumbnailContainer.js'
+import Header from '../Header/Header.js';
+import Movie from '../Movie/Movie.js';
+import movieData from '../movieData.js';
+import '../App/App.scss';
 
 class App extends Component {
   constructor() {
@@ -38,11 +38,11 @@ class App extends Component {
     return (
       <main className='App'>
       <Header displayHome={this.displayHome}/>
-        { !this.state.selectedMovie ? 
+        { !this.state.selectedMovie ?
         <ThumbnailContainer
           movies={this.state.movies}
           displayMovie={this.displayMovie}
-        /> 
+        />
         : <Movie movie={this.state.selectedMovie}/> }
       </main>
     )
