+++ src/Header/Header.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../css/Header.scss'
+import './Header.scss'
 
 const Header = ({ displayHome }) => {
   return (
