+++ src/Thumbnail-Container/ThumbnailContainer.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import Thumbnail from './Thumbnail.js'
-import '../css/ThumbnailContainer.scss'
+import Thumbnail from '../Thumbnail/Thumbnail.js'
+import './ThumbnailContainer.scss'
 
 const ThumbnailContainer = ({ movies, displayMovie }) => {
   const allMovies = movies.map(movie => {
