+++ src/Thumbnail-Container/ThumbnailContainer.scss
