+++ src/Thumbnail/Thumbnail.scss
